Added a trait for EntitySpecificationRepository
