+++ src/EntitySpecificationRepository.php
@@ -4,29 +4,10 @@ namespace Happyr\DoctrineSpecification;
 
 use Doctrine\ORM\EntityRepository;
 
+/**
+ * This class allows you to use a Specification to query entities
+ */
 class EntitySpecificationRepository extends EntityRepository
 {
-    private $alias = 'e';
-
-    /**
-     * Get result when you match with a Specification
-     *
-     * @param Specification   $specification
-     * @param Result\ResultModifier $modifier
-     *
-     * @return mixed
-     */
-    public function match(Specification $specification, Result\ResultModifier $modifier = null)
-    {
-        $qb = $this->createQueryBuilder($this->alias);
-
-        $specification->modify($qb, $this->alias);
-        $query = $qb->where($specification->getFilter($qb, $this->alias))->getQuery();
-
-        if ($modifier instanceof Result\ResultModifier) {
-            $modifier->modify($query);
-        }
-
-        return $query->execute();
-    }
+    use EntitySpecificationRepositoryTrait;
 }
